fix: split profile & history page
